+++ basic/src/tags/components/CreateTagForm.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { createTag } from "wasp/client/operations";
-import { Input } from "../../common/Input";
+import { Input } from "../../components/Input";
 import { ColorRadioButtons } from "./ColorRadioButtons";
 import { generateBrightColor } from "./colors";
 import { TagLabel } from "./TagLabel";
